requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Azure SDKs
azure-ai-documentintelligence==1.0.0
azure-cosmos==4.9.0
azure-storage-blob==12.24.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry==1.6.4
openai>=1.61

# Microsoft Agent Framework (Semantic Kernel)
semantic-kernel[azure]==1.21.1

# Azure AI Foundry (Evaluation + Tracing)
azure-ai-projects==1.0.0b7
azure-ai-evaluation==1.6.0
opentelemetry-sdk==1.29.0

# Document Processing
python-docx==1.1.2
PyPDF2==3.0.1

# Utilities
pydantic>=2.10.4,<2.11
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
websockets==14.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Code Quality
ruff==0.8.4
